Create README.md
This Python code is intended to allow for a means to extract parameters of interest from the mavlink2rest protocol from a BlueROV2 companion computer.
 
The BlueROV2 companion computer publishes all relevant data streams from the BlueROV2 as the following location:
http://xxx.xxx.xxx.xxx/mavlink. 

This script assumes that the BlueROV2 computer is assigned the static IP address of 192.168.3.200, and thus REST data is available at http://192.168.3.200/mavlink. Data of interest are nested beneath this top level hierarchy, and are more fully detailed in the ArduSub documentation. The key parameters of interest to export are the BlueROV2's depth, heading, umbilical turns and altitude. 

Data are extracted using the Python requests module, and are concatenated to make a single, comma seperated data string. The current OS system time from the PC running this script are pre-pended to the comma seperated string. 

The full string is re-broadcast through a TCP server instantiated through this code. At present, this program supports on a single client connection to the data stream, intended to be used by Hypack.